Create iperf2_bounceback_pdf_cdf_generator
This script is intended to generate a pdf and cdf using an output like the one from iperf2 bounceback test, where values are separated by ":", where the left one is the latency (x100us) and the right one is the count.

It is possible to add many distributions on the same plot for comparison. The script will ask for a name of each distribution first (free text) and then the values for that name must be inserted. To finish the script, one must enter 'done'.